requirements.txt
genanki
PyYAML